backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.165.1",
    "crewai-tools>=0.62.3",
    "faker>=37.5.3",
    "fastapi[standard]>=0.116.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.2.14",
    "pandas>=2.3.2",
    "pydantic>=2.11.7",
    "python-ulid>=3.1.0",
    "uuid>=1.30",
]
